+++ client/src/components/common/icons.tsx
@@ -1,0 +1,60 @@
+import React from 'react';
+import { createIcon, Icon } from '@chakra-ui/react';
+
+const RecordIcon = createIcon({
+  displayName: 'RecordIcon',
+  path: (
+    <svg
+    //   //   enable-background="new 0 0 141.732 141.732"
+    //   height="141.732px"
+    //   id="Livello_1"
+    //   version="1.1"
+    //   viewBox="0 0 141.732 141.732"
+    //   width="141.732px"
+    //   //   xml:space="preserve"
+    //   xmlns="http://www.w3.org/2000/svg"
+    //   //   xmlns:xlink="http://www.w3.org/1999/xlink"
+    >
+      <g id="Livello_7">
+        <path d="M91.192,59.623V21.946c0.043-0.534,0.068-1.073,0.068-1.618C91.262,9.102,82.161,0,70.935,0   c-11.229,0-20.33,9.104-20.33,20.328c0,0.545,0.024,1.084,0.066,1.618v37.677c-0.042,0.535-0.066,1.073-0.066,1.621   c0,11.229,9.103,20.326,20.329,20.326c11.227,0,20.327-9.1,20.327-20.326C91.262,60.696,91.235,60.158,91.192,59.623    M120.942,50.219c0-3.084-2.5-5.584-5.584-5.584c-3.082,0-5.582,2.5-5.582,5.584c0,0.047,0.004,0.092,0.008,0.139   c-0.076,21.43-17.471,38.779-38.917,38.779c-21.448,0-38.842-17.353-38.917-38.779c0.002-0.047,0.008-0.092,0.008-0.139   c0-3.084-2.5-5.584-5.584-5.584s-5.585,2.5-5.585,5.584c0,25.714,19.388,46.896,44.339,49.743v28.648H38.78   c-2.96,0-5.358,2.402-5.358,5.361c0,2.961,2.398,5.355,5.358,5.355h26.35h11.61h26.35c2.959,0,5.357-2.396,5.357-5.355   c0-2.959-2.398-5.361-5.357-5.361h-26.35V99.945C101.628,97.039,120.942,75.885,120.942,50.219" />
+      </g>
+      <g id="Livello_1_1_" />
+    </svg>
+  )
+});
+
+const FlagIcon = createIcon({
+  displayName: 'flagIcon',
+  path: (
+    <svg
+      version="1.1"
+      id="Capa_1"
+      xmlns="http://www.w3.org/2000/svg"
+      x="0px"
+      y="0px"
+      viewBox="0 0 206.505 206.505"
+      xmlSpace="preserve"
+    >
+      <g>
+        <path
+          d="M167.069,16.354c-0.426,0.186-43.29,18.671-70.274-4.481c-31.021-26.619-60.722-0.2-61.012,0.072
+		L34.623,13l-0.179,193.494l7.108,0.011l0.115-122.594c5.157-4.091,27.11-19.086,50.498,0.988
+		c30.356,26.047,75.839,6.438,77.765,5.59l2.133-0.938V14.153L167.069,16.354z M164.943,84.824
+		c-8.997,3.389-44.603,14.874-68.148-5.322c-10.375-8.904-20.6-11.871-29.629-11.871c-10.815,0-19.923,4.266-25.489,7.702
+		l0.047-59.101c5.264-4.169,27.142-18.961,50.44,1.034c24.855,21.323,59.863,12.043,72.78,7.516V84.824z"
+        />
+      </g>
+    </svg>
+  )
+});
+
+const CircleIcon = (props: any) => (
+  <Icon viewBox="0 0 200 200" {...props}>
+    <path
+      fill="currentColor"
+      d="M 100, 100 m -75, 0 a 75,75 0 1,0 150,0 a 75,75 0 1,0 -150,0"
+    />
+  </Icon>
+);
+
+export { RecordIcon, FlagIcon, CircleIcon };
